+++ app/javascript/vue/tasks/matrix_row_coder/store/actions/requestMatrixRow.js
@@ -3,7 +3,7 @@ import DescriptorTypes from '../helpers/DescriptorTypes'
 import ComponentNames from '../helpers/ComponentNames'
 import makeEmptyObservationsFor from '../helpers/makeEmptyObservationsFor'
 
-export default function ({commit, state}, args) {
+export default function ({ commit, state }, args) {
   const {
     rowId,
     otuId
@@ -56,6 +56,7 @@ function makeBaseDescriptor (descriptorData) {
 }
 
 const DescriptorTypesToComponentNames = {
+  [DescriptorTypes.FreeText]: ComponentNames.FreeText,
   [DescriptorTypes.Media]: ComponentNames.Media,
   [DescriptorTypes.Qualitative]: ComponentNames.Qualitative,
   [DescriptorTypes.Continuous]: ComponentNames.Continuous,