+++ app/javascript/vue/tasks/matrix_row_coder/store/helpers/makeObservation.js
@@ -1,9 +1,10 @@
 import makeBaseObservation from './makeBaseObservation'
-import makeQualitativeObservation from './makeQualitativeObservation'
 import makeContinuousObservation from './makeContinuousObservation'
+import makeFreeTextObservation from './makeFreeTextObservation'
+import makeMediaObservation from './makeMediaObservation'
 import makePresenceObservation from './makePresenceObservation'
 import makeSampleObservation from './makeSampleObservation'
-import makeMediaObservation from './makeMediaObservation'
+import makeQualitativeObservation from './makeQualitativeObservation'
 import ObservationTypes from './ObservationTypes'
 
 export default function (observationData) {
@@ -17,6 +18,8 @@ export default function (observationData) {
     return makeSampleObservation(observationData)
   } else if (observationData.type === ObservationTypes.Media) {
     return makeMediaObservation(observationData)
+  } else if (observationData.type === ObservationTypes.FreeText) {
+    return makeFreeTextObservation(observationData)
   }
   return makeBaseObservation(observationData)
 }